go to infinit sleep after init on eMMC

+++ scripts/waggle_init.sh
@@ -535,4 +535,12 @@ else
   echo "all looks good" 
 fi
 
+
+if [ "${CURRENT_DEVICE_TYPE}x" == "MMCx" ] ; then
+  echo "Detected MMC, will go to sleep to prevent nodecontroller software from starting"
+  sleep infinity
+  exit 1
+fi
+
+
 rm -f ${pidfile}
